refactor: integerates an interface for yaml file handling
