+++ .github/workflows/go.yml
@@ -27,6 +27,15 @@ jobs:
 
       - name: Build
         run: go build -v ./...
+
+       - name: Test-Export
+        run: |
+          go test -coverpkg=./... -coverprofile=coverage.tmp.txt -covermode=atomic ./...
+          cat coverage.tmp.txt | grep -v -e "generated.go" -e "_gen.go" | grep -v -e "services.pb.go" -e "services_grpc.pb.go" > coverage.txt
+        env:
+          KEPLOY_API_KEY: 81f83aeeedddf453966347dc136c66
+          ENABLE_DEDUP: false
+          ENABLE_TEST_EXPORT: true
         
       - name: Test
         run: |
